changed mealplan price to positive integer

+++ products/models.py
@@ -37,7 +37,7 @@ class Mealplan(models.Model):
     # columns in the table
     title = models.CharField(blank=False, max_length=255)
     description = models.CharField(blank=False, max_length=255)
-    price = models.DecimalField(max_digits=10, decimal_places=3, blank=False)
+    price = models.PositiveIntegerField(blank=False)
     mealplan_type = models.ForeignKey(MealplanType, on_delete=models.CASCADE)
 
     # to make string
